Hide add panel icon if there isn't a layout
